+++ packages/presets/src/theme/recipes/avatar.ts
@@ -3,7 +3,6 @@ import { defineSlotRecipe } from '@pandacss/dev'
 
 export const avatar = defineSlotRecipe({
   className: 'avatar',
-  description: 'An avatar style',
   slots: avatarAnatomy.keys(),
   base: {
     root: {