+++ packages/presets/src/theme/recipes/badge.ts
@@ -2,7 +2,6 @@ import { defineRecipe } from '@pandacss/dev'
 
 export const badge = defineRecipe({
   className: 'badge',
-  description: 'A badge style',
   base: {
     borderRadius: 'full',
     fontWeight: 'medium',
@@ -26,6 +25,9 @@ export const badge = defineRecipe({
         borderColor: 'border.default',
         borderWidth: '1px',
         color: 'fg.emphasized',
+        '& svg': {
+          color: 'fg.muted',
+        },
       },
       outline: {
         borderWidth: '2px',
@@ -33,9 +35,36 @@ export const badge = defineRecipe({
       },
     },
     size: {
-      sm: { textStyle: 'xs', px: '2', h: '5', gap: '1' },
-      md: { textStyle: 'sm', px: '2.5', h: '6', gap: '1.5' },
-      lg: { textStyle: 'sm', px: '3', h: '7', gap: '1.5' },
+      sm: {
+        textStyle: 'xs',
+        px: '2',
+        h: '5',
+        gap: '1',
+        '& svg': {
+          width: '3',
+          height: '3',
+        },
+      },
+      md: {
+        textStyle: 'sm',
+        px: '2.5',
+        h: '6',
+        gap: '1.5',
+        '& svg': {
+          width: '4',
+          height: '4',
+        },
+      },
+      lg: {
+        textStyle: 'sm',
+        px: '3',
+        h: '7',
+        gap: '1.5',
+        '& svg': {
+          width: '4',
+          height: '4',
+        },
+      },
     },
   },
 })