+++ packages/presets/src/theme/recipes/button.ts
@@ -2,7 +2,6 @@ import { defineRecipe } from '@pandacss/dev'
 
 export const button = defineRecipe({
   className: 'button',
-  description: 'A button style',
   base: {
     alignItems: 'center',
     appearance: 'none',
@@ -22,12 +21,12 @@ export const button = defineRecipe({
     whiteSpace: 'nowrap',
   },
   defaultVariants: {
-    variant: 'primary',
+    variant: 'solid',
     size: 'md',
   },
   variants: {
     variant: {
-      primary: {
+      solid: {
         background: 'accent.default',
         color: 'accent.fg',
         _hover: {
@@ -46,7 +45,7 @@ export const button = defineRecipe({
           },
         },
       },
-      secondary: {
+      outline: {
         borderWidth: '1px',
         borderColor: 'border.emphasized',
         background: 'bg.default',
@@ -81,7 +80,7 @@ export const button = defineRecipe({
           outlineColor: 'border.outline',
         },
       },
-      tertiary: {
+      ghost: {
         color: 'fg.emphasized',
         _hover: {
           color: 'fg.default',