+++ packages/presets/src/theme/recipes/index.ts
@@ -14,6 +14,7 @@ import { datePicker } from './date-picker'
 import { dialog } from './dialog'
 import { drawer } from './drawer'
 import { editable } from './editable'
+import { heading } from './heading'
 import { hoverCard } from './hover-card'
 import { icon } from './icon'
 import { input } from './input'
@@ -33,6 +34,7 @@ import { select } from './select'
 import { slider } from './slider'
 import { splitter } from './splitter'
 import { switchRecipe } from './switch'
+import { table } from './table'
 import { tabs } from './tabs'
 import { tagsInput } from './tags-input'
 import { textarea } from './textarea'
@@ -41,11 +43,11 @@ import { toggleGroup } from './toggle-group'
 import { tooltip } from './tooltip'
 
 export const recipes = {
-  alert,
   article,
   badge,
   button,
   code,
+  heading,
   icon,
   input,
   label,
@@ -55,6 +57,7 @@ export const recipes = {
 
 export const slotRecipes = {
   accordion,
+  alert,
   avatar,
   card,
   carousel,
@@ -81,6 +84,7 @@ export const slotRecipes = {
   splitter,
   switchRecipe,
   tabs,
+  table,
   tagsInput,
   toast,
   toggleGroup,