+++ website/src/components/demos/badge.demo.tsx
@@ -1,0 +1,5 @@
+import { Badge, type BadgeProps } from '~/components/ui/badge'
+
+export const Demo = (props: BadgeProps) => {
+  return <Badge {...props}>Park UI</Badge>
+}