+++ website/src/components/ui/button.tsx
@@ -1,8 +1,6 @@
-'use client' // TODO this should not be necessary
-import { ark } from '@ark-ui/react'
-import type { ComponentPropsWithoutRef } from 'react'
+import { ark, type HTMLArkProps } from '@ark-ui/react'
 import { styled } from 'styled-system/jsx'
 import { button, type ButtonVariantProps } from 'styled-system/recipes'
 
-export type ButtonProps = ButtonVariantProps & ComponentPropsWithoutRef<typeof ark.button>
+export type ButtonProps = ButtonVariantProps & HTMLArkProps<'button'>
 export const Button = styled(ark.button, button)